feat: Add use-clear-user-data hook
This commit adds a new hook called use-clear-user-data. This hook provides a function called clearData that can be used to clear all locally stored user data. The hook uses the useSetAtom function from the jotai library to update the relevant atoms.

The clearData function resets the following atoms to their initial values:
- stxAddressAtom
- accountDataAtom
- signatureMsgAtom
- signatureDataAtom
- sentDustToggleAtom
- insufficientBalanceToggleAtom

This hook can be used in the clear-data.tsx and sign-in.tsx components to clear user data when the "Clear Data" button is clicked or when the user signs in with a different STX address.

Note: The useClearUserData hook is a new file added to the src/hooks directory.